requirements.txt
fastapi
uvicorn[standard]
paramiko
python-dotenv